Fixing error introduced in commit: 16cc8aeeda481d.
Initially the popover view controller was created
when the BKSplitViewController class was
initialized and the "master" view controller was
assigned to the popover controllers
contentViewController property when the array of
view controllers (viewControllers) was updated.
This meant that the "master" view controller was
assigned to the popover view controller when
displaying both views in portrait orientations.